Refactor: move NightWrite class info out of it's runner file
