+++ message.txt
@@ -1,3 +1,1 @@
-hello world
-hello world
-hello wor
+nothing
