+++ spec/night_writer_class_spec.rb
@@ -1,4 +1,5 @@
 require './lib/night_writer'
+require './lib/night_writer_class'
 require 'tempfile'
 
 RSpec.describe NightWriter do
